Extracted venues view controller into its own class
