+++ Lunchify/Base.lproj/Main.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="10117" systemVersion="15G31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="1Qc-im-KrP">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="10116" systemVersion="15G31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="1Qc-im-KrP">
     <dependencies>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="10085"/>
         <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
@@ -102,6 +102,10 @@
                             <action selector="refreshVenues:" destination="VJN-r5-A4U" eventType="valueChanged" id="vi5-qh-U1t"/>
                         </connections>
                     </refreshControl>
+                    <connections>
+                        <outlet property="refreshControl" destination="mxZ-ET-xgt" id="Z4I-py-wHr"/>
+                        <outlet property="tableView" destination="LDt-Sj-kD9" id="lkL-KS-YOY"/>
+                    </connections>
                 </tableViewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Nr4-Xg-Nrg" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
